refactor(folder): 🧹 restructure folder structure instead on flat files
